--- src/main/java/com/dia/delievery/common/advice/GolbalControllerAdvice.java
@@ -1,8 +1,0 @@
-package com.dia.delievery.common.advice;
-
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-@RestControllerAdvice
-public class GolbalControllerAdvice {
-
-}
