--- src/main/java/com/dia/delievery/common/security/UserDetailsImpl.java
@@ -1,70 +1,0 @@
-package com.dia.delievery.common.security;
-
-
-import com.dia.delievery.user.UserRoleEnum;
-import com.dia.delievery.user.entity.Users;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-
-import java.util.ArrayList;
-import java.util.Collection;
-
-public class UserDetailsImpl implements UserDetails {
-
-    private final Users users;
-
-    public UserDetailsImpl(Users users) {
-        this.users = users;
-    }
-
-    public Users getUser() {
-        return users;
-    }
-
-    public String getRole(){
-        return users.getRole().toString();
-    }
-
-    @Override
-    public String getPassword() {
-        return users.getPassword();
-    }
-
-    @Override
-    public String getUsername() {
-        return users.getUsername();
-    }
-
-    @Override
-    public Collection<? extends GrantedAuthority> getAuthorities() {
-        UserRoleEnum role = users.getRole();
-        String authority = role.getAuthority();
-
-        SimpleGrantedAuthority simpleGrantedAuthority = new SimpleGrantedAuthority(authority);
-        Collection<GrantedAuthority> authorities = new ArrayList<>();
-        authorities.add(simpleGrantedAuthority);
-
-        return authorities;
-    }
-
-    @Override
-    public boolean isAccountNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isAccountNonLocked() {
-        return true;
-    }
-
-    @Override
-    public boolean isCredentialsNonExpired() {
-        return true;
-    }
-
-    @Override
-    public boolean isEnabled() {
-        return true;
-    }
-}
