--- src/main/java/com/dia/delievery/common/security/UserDetailsServiceImpl.java
@@ -1,27 +1,0 @@
-package com.dia.delievery.common.security;
-
-
-import com.dia.delievery.user.entity.Users;
-import com.dia.delievery.user.repository.UserRepository;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserDetailsServiceImpl implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    public UserDetailsServiceImpl(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        Users users = userRepository.findByUsername(username)
-                .orElseThrow(() -> new UsernameNotFoundException("Not Found " + username));
-
-        return new UserDetailsImpl(users);
-    }
-}
