--- src/main/java/com/dia/delievery/review/entity/Reviews.java
@@ -1,28 +1,0 @@
-package com.dia.delievery.review.entity;
-
-import com.dia.delievery.store.entity.Stores;
-import com.dia.delievery.user.entity.Users;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-@Getter
-public class Reviews {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @Column
-    private String content;
-    @Column
-    private String imageUrl;
-    @Column
-    private double rate;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Users users;
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Stores stores;
-}
