--- src/main/java/com/dia/delievery/user/UserRoleEnum.java
@@ -1,23 +1,0 @@
-package com.dia.delievery.user;
-
-public enum UserRoleEnum {
-    USER(Authority.USER),
-    OWNER(Authority.OWNER),
-    ADMIN(Authority.ADMIN);
-
-    private final String authority;
-
-    UserRoleEnum(String authority) {
-        this.authority = authority;
-    }
-
-    public String getAuthority() {
-        return this.authority;
-    }
-
-    public static class Authority {
-        public static final String USER = "ROLE_USER";
-        public static final String OWNER = "ROLE_OWNER";
-        public static final String ADMIN = "ROLE_ADMIN";
-    }
-}
