--- src/main/java/com/dia/delievery/user/dto/LoginRequestDto.java
@@ -1,12 +1,0 @@
-package com.dia.delievery.user.dto;
-
-import lombok.Getter;
-import lombok.Setter;
-
-@Getter
-@Setter
-public class LoginRequestDto {
-    private String username;
-    private String password;
-    //푸시
-}
