--- src/main/java/com/dia/delievery/user/repository/UserRepository.java
@@ -1,10 +1,0 @@
-package com.dia.delievery.user.repository;
-
-import com.dia.delievery.user.entity.Users;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<Users, Long> {
-    Optional<Users> findByUsername(String username);
-}
