--- src/main/java/com/dia/delievery/userscrapstore/entity/UserScrapStore.java
@@ -1,21 +1,0 @@
-package com.dia.delievery.userscrapstore.entity;
-
-import com.dia.delievery.store.entity.Stores;
-import com.dia.delievery.user.entity.Users;
-import jakarta.persistence.*;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Entity
-@NoArgsConstructor
-@Getter
-public class UserScrapStore {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Users users;
-    @ManyToOne(fetch = FetchType.LAZY)
-    private Stores stores;
-}
