+++ src/main/java/com/dia/delivery/common/advice/GolbalControllerAdvice.java
@@ -1,0 +1,8 @@
+package com.dia.delivery.common.advice;
+
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+@RestControllerAdvice
+public class GolbalControllerAdvice {
+
+}
