+++ src/main/java/com/dia/delivery/product/entity/Products.java
@@ -1,0 +1,36 @@
+package com.dia.delivery.product.entity;
+
+import com.dia.delivery.store.entity.Stores;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@Getter
+@NoArgsConstructor
+public class Products {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(nullable = false)
+    private String productName;
+
+    @Column
+    private String imageUrl;
+
+    @Column(nullable = false)
+    private String link;
+
+    @Column(nullable = false)
+    private int price;
+
+    @Column(nullable = false)
+    private String description;
+
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(nullable = false)
+    private Stores stores;
+}
