+++ src/main/java/com/dia/delivery/review/entity/Reviews.java
@@ -1,0 +1,28 @@
+package com.dia.delivery.review.entity;
+
+import com.dia.delivery.store.entity.Stores;
+import com.dia.delivery.user.entity.Users;
+import jakarta.persistence.*;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Entity
+@NoArgsConstructor
+@Getter
+public class Reviews {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column
+    private String content;
+    @Column
+    private String imageUrl;
+    @Column
+    private double rate;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Users users;
+    @ManyToOne(fetch = FetchType.LAZY)
+    private Stores stores;
+}
