+++ src/main/java/com/dia/delivery/user/UserRoleEnum.java
@@ -1,0 +1,23 @@
+package com.dia.delivery.user;
+
+public enum UserRoleEnum {
+    USER(Authority.USER),
+    OWNER(Authority.OWNER),
+    ADMIN(Authority.ADMIN);
+
+    private final String authority;
+
+    UserRoleEnum(String authority) {
+        this.authority = authority;
+    }
+
+    public String getAuthority() {
+        return this.authority;
+    }
+
+    public static class Authority {
+        public static final String USER = "ROLE_USER";
+        public static final String OWNER = "ROLE_OWNER";
+        public static final String ADMIN = "ROLE_ADMIN";
+    }
+}
