+++ src/main/java/com/dia/delivery/user/dto/LoginRequestDto.java
@@ -1,0 +1,12 @@
+package com.dia.delivery.user.dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class LoginRequestDto {
+    private String username;
+    private String password;
+    //푸시
+}
