+++ src/main/java/com/dia/delivery/user/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.dia.delivery.user.repository;
+
+import com.dia.delivery.user.entity.Users;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<Users, Long> {
+    Optional<Users> findByUsername(String username);
+}
