--- src/test/java/com/dia/delievery/DelieveryApplicationTests.java
@@ -1,13 +1,0 @@
-package com.dia.delievery;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class DelieveryApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
