+++ src/test/java/com/dia/delivery/DelieveryApplicationTests.java
@@ -1,0 +1,13 @@
+package com.dia.delivery;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class DelieveryApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
